+++ activiti-cloud-audit-mongo-dependencies/pom.xml
@@ -18,7 +18,7 @@
         <artifactId>activiti-cloud-dependencies-parent</artifactId>
         <version>${activiti-cloud-build.version}</version>
         <scope>import</scope>
-        <type>pom</type>       
+        <type>pom</type>
       </dependency>
       <dependency>
         <groupId>org.activiti.cloud.common</groupId>
@@ -29,16 +29,18 @@
       </dependency>
       <dependency>
         <groupId>org.activiti.cloud.audit</groupId>
-        <artifactId>activiti-cloud-audit-service-mongo-parent</artifactId>
-        <version>${activiti-cloud-audit-mongo.version}</version>
+        <artifactId>activiti-cloud-audit-dependencies</artifactId>
+        <version>${activiti-cloud-audit.version}</version>
+        <type>pom</type>
+        <scope>import</scope>
       </dependency>
       <dependency>
-        <groupId>org.activiti.cloud.audit</groupId>
+        <groupId>org.activiti.cloud.audit.mongo</groupId>
         <artifactId>activiti-cloud-services-audit-mongo</artifactId>
         <version>${activiti-cloud-audit-mongo.version}</version>
       </dependency>
       <dependency>
-        <groupId>org.activiti.cloud.audit</groupId>
+        <groupId>org.activiti.cloud.audit.mongo</groupId>
         <artifactId>activiti-cloud-starter-audit-mongo</artifactId>
         <version>${activiti-cloud-audit-mongo.version}</version>
       </dependency>
